+++ Circuit Breaker/Assets/Circuit Breaker/Scripts/CircuitComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 070373f8e32929d4bbf048693273c574
+guid: 1fde7209492cc9e49bac70c9752bfa3d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
